requirements.txt
torch
torchvision
torchaudio
numpy
pandas
pillow
matplotlib
jupyter
ipykernel
fastapi
uvicorn
python-multipart
streamlit
